+++ wiki/Changelog.md
@@ -4,6 +4,7 @@ Terminal features
   * Support multiple controls in OSC 4, 5, 10..19, 104, 105 (#1038, xterm).
 
 Window handling
+  * Click-open link still after moving the mouse over the link area (#1039).
   * Keep hotkey-started window in taskbar (#1035).
   * Override font zooming also with Ctrl, to support FancyZones (#487).